implement an extended periodic packet mechanism
With this mechanism, it is possible to configure the unit to send
different packets at different periods. Packets are multiplexed in
a single "ep" packet and must be demultiplexed on the receiving side.

The mechanism is configured through a byte array in the configuration
structure, where each entry is the period of the corresponding packet.