fix: tune the standard deviation for the mag yaw estimates
The current variance is 0.1, which means ~15 deg standard deviation.
Reduce it to 0.01. Note that the 0.01 value is what happens when the
turn switch is TRUE, which is probably what happens when standing still
(that is, still in the Marina), where I observed a static 0.5 degree
error w.r.t. RTK.

Also, reestablish the x2 factor when roll and pitch are high (the
justification for those makes sense)