chore: rework heading handling
This partially removes code necessary to use GPS velocity
as heading source. It cleans up our case, that is the one
where we want to seamlessly switch between RTK heading, mag
heading and mag heading with declination.

The main issue with the GNSS heading is that it is not available
*at all* at zero velocities, and that it needs time to switch
(some validation / integration times). This makes it a lot harder
to implement, and since we don't need it, I removed (most of) it.

The heading source is now chosen within the update step. The
correction of mag yaw measurement by declination is done there
as well to detect the transition from mag and mag-with-decl